+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Limit number of mentions and links in remote posts.
+
 ## [1.19.0] - 2023-03-30
 
 ### Added